# swagger/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dating App for Dealls",
        "title": "Dating App",
        "contact": {
            "name": "@usagifm",
            "url": "https://example.id"
        },
        "version": "1.0"
    },
    "host": "dating-app.taktix.co.id",
    "basePath": "/",
    "paths": {
...
        "/api/v1/dating/package/buy": {
            "post": {
                "description": "Purchase a package",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Package"
                ],
                "summary": "Buy a package",
                "parameters": [
                    {
                        "description": "Buy package request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.BuyPackageRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/v1/dating/preference": {
            "get": {
                "description": "Retrieve the user's preference settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user preference",
                "responses": {}
            },
            "put": {
                "description": "Update the user's preference settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user preference",
                "parameters": [
                    {
                        "description": "Update user preference request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.UpdateUserPreferenceRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/dating/swipe": {
            "post": {
                "description": "Swipe a user and check if it's a match",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Swipe user",
                "parameters": [
                    {
                        "description": "Swipe request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.SwipeRequest"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "contract.BuyPackageRequest": {
            "type": "object",
            "required": [
                "package_id"
            ],
            "properties": {
                "package_id": {
                    "type": "integer"
                }
            }
        },
...
        "contract.SwipeRequest": {
            "type": "object",
            "properties": {
                "swipe_type": {
                    "type": "string"
                },
                "swiped_id": {
                    "type": "integer"
                }
            }
        },
        "contract.UpdateUserPreferenceRequest": {
            "type": "object",
            "properties": {
                "max_age": {
                    "type": "integer"
                },
                "min_age": {
                    "type": "integer"
                },
                "preferred_gender": {
                    "type": "string"
                }
            }
        }
    }
}